Add files via upload
Working with dates in Power Query is interesting. I did this project and also learned some new things that I was not familiar with. I had to convert text to date, date or time elements from date/time values. This project is where the data entry teams typed into the products' released dates five different date format. 7/1/2018, 2018-01-07,  7.1,2018, Jul 1, 2018, 1 July, 2018. The dates had to be in the same regional locale and that was what I did the first. The second project is when I noticed that the dates were from different locale which is one date was US and the other was UK. I split by delimiter to separate the columns and then a custom column and a condition. Check it out please and give me a feedback on what I should have done better. Thank you.